requirements.txt
asgiref==3.8.1
Django==5.1.5
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
Faker==35.0.0
gunicorn==23.0.0
loguru==0.7.3
packaging==24.2
psycopg2-binary==2.9.10
PyJWT==2.10.1
python-dateutil==2.9.0.post0
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.12.2
